+++ example/spec/helpers/testUtils.js
@@ -53,6 +53,7 @@ function loadConfig() {
 
   config.test_configs.buckets = config.buckets;
   config.test_configs.deployment = config.deployment;
+  config.test_configs.cmr = config.cmr;
 
   return config.test_configs;
 }