+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Added `@cumulus/common/util#isUndefined()`
   - Added `@cumulus/common/util#negate()`
 
+- CUMULUS-1151 Compare the granule holdings in CMR with Cumulus' internal data store
+
 ### Changed
 
 - **CUMULUS-672**